kéo thả tháp khi tháp có kích thước lớn hơn tile map gây lỗi vị trí
